# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PwTokenInput {
  /** @minLength 1 */
  token: string;
}

export interface PwOtpInput {
  /** @minLength 3 */
  phone: string;
  /** @minLength 1 */
  countryCode: string;
}

export interface PwOtpExchangeInput {
  /** @minLength 3 */
  phone: string;
  /** @minLength 1 */
  otp: string;
}

export interface PwAuthResult {
  valid: boolean;
  /** @nullable */
  errorMessage?: string | null;
}

export interface PwActionResult {
  success: boolean;
  /** @nullable */
  errorMessage?: string | null;
}

export interface PwTokenResult {
  success: boolean;
  /** @nullable */
  accessToken?: string | null;
  /** @nullable */
  expiresIn?: number | null;
  /** @nullable */
  errorMessage?: string | null;
}

export interface PwBatch {
  /** @nullable */
  id?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  expiryDate?: string | null;
}

export interface PwSubject {
  /** @nullable */
  id?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  lectureCount?: number | null;
}

export interface PwTopic {
  /** @nullable */
  id?: string | null;
  name: string;
  slug: string;
}

export interface PwAttachment {
  /** @nullable */
  id?: string | null;
  name: string;
  url: string;
}

export interface PwResourceGroup {
  /** @nullable */
  topic?: string | null;
  attachments: PwAttachment[];
}

export interface PwAnnouncement {
  /** @nullable */
  id?: string | null;
  announcement: string;
  /** @nullable */
  scheduleTime?: string | null;
  attachment: PwAttachment | null;
}

/**
 * Invalid request
 */
export type BadRequestResponse = PwActionResult;

/**
 * Missing or invalid PW session token
 */
export type UnauthorizedResponse = PwActionResult;