# lib/api-zod/src/generated/types/pwActionResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface PwActionResult {
  success: boolean;
  /** @nullable */
  errorMessage?: string | null;
}