# lib/api-zod/src/generated/types/pwAnnouncement.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { PwAttachment } from './pwAttachment';

export interface PwAnnouncement {
  /** @nullable */
  id?: string | null;
  announcement: string;
  /** @nullable */
  scheduleTime?: string | null;
  attachment: PwAttachment | null;
}