# lib/api-zod/src/generated/types/pwAuthResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface PwAuthResult {
  valid: boolean;
  /** @nullable */
  errorMessage?: string | null;
}