# lib/api-zod/src/generated/types/pwBatch.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface PwBatch {
  /** @nullable */
  id?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  expiryDate?: string | null;
}