# lib/api-zod/src/generated/types/pwOtpExchangeInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface PwOtpExchangeInput {
  /** @minLength 3 */
  phone: string;
  /** @minLength 1 */
  otp: string;
}