# lib/api-zod/src/generated/types/pwResourceGroup.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */
import type { PwAttachment } from './pwAttachment';

export interface PwResourceGroup {
  /** @nullable */
  topic?: string | null;
  attachments: PwAttachment[];
}