# lib/api-zod/src/generated/types/pwSubject.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface PwSubject {
  /** @nullable */
  id?: string | null;
  name: string;
  slug: string;
  /** @nullable */
  lectureCount?: number | null;
}