# lib/api-zod/src/generated/types/pwTokenResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface PwTokenResult {
  success: boolean;
  /** @nullable */
  accessToken?: string | null;
  /** @nullable */
  expiresIn?: number | null;
  /** @nullable */
  errorMessage?: string | null;
}