# lib/api-zod/src/generated/types/pwTopic.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * PW Study Vault API
 * OpenAPI spec version: 0.1.0
 */

export interface PwTopic {
  /** @nullable */
  id?: string | null;
  name: string;
  slug: string;
}